+++ app/controllers/local_authorities_controller.rb
@@ -1,0 +1,9 @@
+class LocalAuthoritiesController < ApplicationController
+  def index
+    @local_authorities = LocalAuthority.all.order(:name)
+  end
+
+  def show
+    @local_authority = LocalAuthority.find(params[:id])
+  end
+end