+++ app/helpers/local_authority_helper.rb
@@ -1,0 +1,5 @@
+module LocalAuthorityHelper
+  def address_markup(address)
+    address.compact_blank.join("<br/>")
+  end
+end